added tar downloader and restructured folders
